Fix dynakube always being set to error phase
